# Conflicts:
#	config/model/src/main/java/com/indracompany/sofia2/config/model/DataModel.java
#	config/services/src/main/java/com/indracompany/sofia2/config/services/deletion/EntityDeletionServiceImpl.java
#	config/services/src/main/java/com/indracompany/sofia2/config/services/ontology/OntologyServiceImpl.java

+++ config/services/src/main/java/com/indracompany/sofia2/config/services/deletion/EntityDeletionServiceImpl.java
@@ -20,6 +20,7 @@ import org.springframework.stereotype.Service;
 
 import com.indracompany.sofia2.config.model.Ontology;
 import com.indracompany.sofia2.config.model.TwitterListening;
+import com.indracompany.sofia2.config.model.User;
 import com.indracompany.sofia2.config.repository.ClientPlatformOntologyRepository;
 import com.indracompany.sofia2.config.repository.DeviceSimulationRepository;
 import com.indracompany.sofia2.config.repository.OntologyEmulatorRepository;
@@ -27,11 +28,12 @@ import com.indracompany.sofia2.config.repository.OntologyRepository;
 import com.indracompany.sofia2.config.repository.OntologyUserAccessRepository;
 import com.indracompany.sofia2.config.repository.TwitterListeningRepository;
 import com.indracompany.sofia2.config.services.exceptions.OntologyServiceException;
-
+import com.indracompany.sofia2.config.services.ontology.OntologyService;
+import com.indracompany.sofia2.config.services.user.UserService;
 
 @Service
-public class EntityDeletionServiceImpl implements EntityDeletionService{
-	
+public class EntityDeletionServiceImpl implements EntityDeletionService {
+
 	@Autowired
 	private OntologyRepository ontologyRepository;
 	@Autowired
@@ -43,41 +45,57 @@ public class EntityDeletionServiceImpl implements EntityDeletionService{
 	@Autowired
 	private TwitterListeningRepository twitterListeningRepository;
 	@Autowired
+
 	private DeviceSimulationRepository deviceSimulationRepository;
-	
-	
-	
-	
+	private OntologyService ontologyService;
+	@Autowired
+	private UserService userService;
+
 	@Override
 	@Transactional
-	public void deleteOntology(String id) {
-		Ontology ontology = this.ontologyRepository.findById(id);
-		try{
-			if(this.clientPlatformOntologyRepository.findByOntology(ontology) != null) {
-				this.clientPlatformOntologyRepository.deleteByOntology(ontology);
+	public void deleteOntology(String id, String userId) {
+
+		try {
+			User user = userService.getUser(userId);
+			Ontology ontology = ontologyService.getOntologyById(id, userId);
+			if (ontologyService.hasUserPermisionForChangeOntology(user, ontology)) {
+				if (this.clientPlatformOntologyRepository.findByOntology(ontology) != null) {
+					this.clientPlatformOntologyRepository.deleteByOntology(ontology);
+				}
+				if (this.ontologyEmulatorRepository.findByOntology(ontology) != null) {
+					this.ontologyEmulatorRepository.deleteByOntology(ontology);
+				}
+				if (this.ontologyUserAccessRepository.findByOntology(ontology) != null) {
+					this.ontologyUserAccessRepository.deleteByOntology(ontology);
+				}
+				if (this.twitterListeningRepository.findByOntology(ontology) != null) {
+					this.twitterListeningRepository.deleteByOntology(ontology);
+				}
+				if (this.ontologyUserAccessRepository.findByOntology(ontology) != null) {
+					this.ontologyUserAccessRepository.deleteByOntology(ontology);
+				}
+				if (this.twitterListeningRepository.findByOntology(ontology) != null) {
+					this.twitterListeningRepository.deleteByOntology(ontology);
+				}
+				if (this.deviceSimulationRepository.findByOntology(ontology) != null) {
+					this.deviceSimulationRepository.deleteByOntology(ontology);
+				}
+				this.ontologyRepository.deleteById(id);
+
+				this.ontologyRepository.deleteById(id);
+			} else {
+				throw new OntologyServiceException("Couldn't delete ontology");
 			}
-			if(this.ontologyEmulatorRepository.findByOntology(ontology) != null) {
-				this.ontologyEmulatorRepository.deleteByOntology(ontology);
-			}
-			if(this.ontologyUserAccessRepository.findByOntology(ontology) != null) {
-				this.ontologyUserAccessRepository.deleteByOntology(ontology);
-			}
-			if(this.twitterListeningRepository.findByOntology(ontology) != null) {
-				this.twitterListeningRepository.deleteByOntology(ontology);
-			}
-			if(this.deviceSimulationRepository.findByOntology(ontology) != null)
-			{
-				this.deviceSimulationRepository.deleteByOntology(ontology);
-			}
-			this.ontologyRepository.deleteById(id);
-		}catch(Exception e){
-			throw new OntologyServiceException("Couldn't delete ontology");
+		} catch (Exception e) {
+			throw new OntologyServiceException("Couldn't delete ontology", e);
 		}
+
 	}
-	
+
 	@Override
 	@Transactional
 	public void deleteTwitterListening(TwitterListening twitterListening) {
 		this.twitterListeningRepository.deleteById(twitterListening.getId());
 	}
+
 }
